backend/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
Flask-Bcrypt
Flask-JWT-Extended
python-dotenv
psycopg2-binary
pymysql
cryptography
